adding contextual actions (CAB) when selecting recipes to generate shopping list of total quantity of ingredients

+++ app/src/main/java/app/kitsu/mercerecetas/database/RecipeIngredientQuantity.kt
@@ -18,7 +18,7 @@ data class RecipeIngredientQuantity (
 
 
     @ColumnInfo(name = "ingredient_quantity")
-    var ingredientQtty: Int = 0,
+    var ingredientQtty: Float = 0f,
 
     @Embedded(prefix = "ing_") val ingredient: Ingredient?
